requirements.txt
python-dotenv==1.0.0
openai==1.3.0
pyyaml==6.0.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0